refactor: extract job tool registrations from MCP server.ts
Move 7 job-related MCP tool registrations (job_complete, job_failed,
job_needs_input, job_log, list_agents, list_recent_persona_reviews,
list_recent_feedback) into a dedicated job-tools.ts module. Reduces
server.ts from 2101 to 1861 lines with no behavior change.
